move test config as function
