+++ packages/rum/karma.conf.js
@@ -23,21 +23,22 @@
  *
  */
 
-var karmaUtils = require('../../dev-utils/karma.js')
+const { baseConfig, prepareConfig } = require('../../dev-utils/karma.js')
+const { getConfig } = require('../../dev-utils/test')
 
 module.exports = function (config) {
   // temporarily set firefox version to 44 due to apm-server#676 issue
-  karmaUtils.baseConfig.customLaunchers.SL_FIREFOX.version = '44'
-  karmaUtils.baseConfig.browserNoActivityTimeout = 120000
+  baseConfig.customLaunchers.SL_FIREFOX.version = '44'
+  baseConfig.browserNoActivityTimeout = 120000
 
-  config.set(karmaUtils.baseConfig)
-  var testConfig = require('../../test.config')
+  config.set(baseConfig)
+  var testConfig = getConfig()
   var customeConfig = {
     globalConfigs: testConfig,
     testConfig: testConfig.env
   }
   console.log('customeConfig:', customeConfig)
   config.set(customeConfig)
-  var cfg = karmaUtils.prepareConfig(config)
+  var cfg = prepareConfig(config)
   config.set(cfg)
 }
